requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf_yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
